+++ packages/soundcloud/src/lib/soundcloud-auto.ts
@@ -1,0 +1,3 @@
+import { register } from './soundcloud'
+
+register()